Sessions - extract view model and renderer
