+++ src/vs/sessions/contrib/changes/browser/changesView.ts
@@ -7,19 +7,15 @@ import './media/changesView.css';
 import * as dom from '../../../../base/browser/dom.js';
 import { renderLabelWithIcons } from '../../../../base/browser/ui/iconLabel/iconLabels.js';
 import { IListVirtualDelegate } from '../../../../base/browser/ui/list/list.js';
-import { ICompressedTreeElement, ICompressedTreeNode } from '../../../../base/browser/ui/tree/compressedObjectTreeModel.js';
-import { ICompressibleTreeRenderer } from '../../../../base/browser/ui/tree/objectTree.js';
-import { IObjectTreeElement, ITreeNode } from '../../../../base/browser/ui/tree/tree.js';
+import { IObjectTreeElement } from '../../../../base/browser/ui/tree/tree.js';
 import { ActionRunner, IAction } from '../../../../base/common/actions.js';
 import { Codicon } from '../../../../base/common/codicons.js';
-import { Iterable } from '../../../../base/common/iterator.js';
-import { Disposable, DisposableStore, IDisposable } from '../../../../base/common/lifecycle.js';
+import { DisposableStore, IDisposable } from '../../../../base/common/lifecycle.js';
 import { Event } from '../../../../base/common/event.js';
-import { autorun, constObservable, derived, derivedObservableWithCache, derivedOpts, IObservable, IObservableWithChange, ISettableObservable, ObservablePromise, observableSignalFromEvent, observableValue, runOnChange } from '../../../../base/common/observable.js';
+import { autorun, derived, IObservable } from '../../../../base/common/observable.js';
 import { basename } from '../../../../base/common/path.js';
-import { IResourceNode, ResourceTree } from '../../../../base/common/resourceTree.js';
 import { ProgressBar } from '../../../../base/browser/ui/progressbar/progressbar.js';
-import { dirname, extUriBiasedIgnorePathCase, isEqual, relativePath } from '../../../../base/common/resources.js';
+import { isEqual } from '../../../../base/common/resources.js';
 import { ThemeIcon } from '../../../../base/common/themables.js';
 import { URI } from '../../../../base/common/uri.js';
 import { localize, localize2 } from '../../../../nls.js';
@@ -32,7 +28,6 @@ import { IActionWidgetDropdownActionProvider } from '../../../../platform/action
 import { IConfigurationService } from '../../../../platform/configuration/common/configuration.js';
 import { IContextKey, IContextKeyService } from '../../../../platform/contextkey/common/contextkey.js';
 import { IContextMenuService } from '../../../../platform/contextview/browser/contextView.js';
-import { FileKind } from '../../../../platform/files/common/files.js';
 import { IHoverService } from '../../../../platform/hover/browser/hover.js';
 import { IInstantiationService, ServicesAccessor } from '../../../../platform/instantiation/common/instantiation.js';
 import { IKeybindingService } from '../../../../platform/keybinding/common/keybinding.js';
@@ -42,31 +37,26 @@ import { ILogService } from '../../../../platform/log/common/log.js';
 import { bindContextKey } from '../../../../platform/observable/common/platformObservableUtils.js';
 import { IOpenerService } from '../../../../platform/opener/common/opener.js';
 import { ServiceCollection } from '../../../../platform/instantiation/common/serviceCollection.js';
-import { IStorageService, StorageScope, StorageTarget } from '../../../../platform/storage/common/storage.js';
+import { IStorageService } from '../../../../platform/storage/common/storage.js';
 import { ITelemetryService } from '../../../../platform/telemetry/common/telemetry.js';
 import { IThemeService } from '../../../../platform/theme/common/themeService.js';
 import { defaultProgressBarStyles } from '../../../../platform/theme/browser/defaultStyles.js';
 import { IWorkspaceContextService } from '../../../../platform/workspace/common/workspace.js';
 import { fillEditorsDragData } from '../../../../workbench/browser/dnd.js';
-import { IResourceLabel, ResourceLabels } from '../../../../workbench/browser/labels.js';
+import { ResourceLabels } from '../../../../workbench/browser/labels.js';
 import { ViewPane, IViewPaneOptions, ViewAction } from '../../../../workbench/browser/parts/views/viewPane.js';
 import { ViewPaneContainer } from '../../../../workbench/browser/parts/views/viewPaneContainer.js';
 import { IViewDescriptorService } from '../../../../workbench/common/views.js';
 import { CHAT_CATEGORY } from '../../../../workbench/contrib/chat/browser/actions/chatActions.js';
 import { IAgentSessionsService } from '../../../../workbench/contrib/chat/browser/agentSessions/agentSessionsService.js';
 import { ChatContextKeys } from '../../../../workbench/contrib/chat/common/actions/chatContextKeys.js';
-import { IChatSessionFileChange, IChatSessionFileChange2, isIChatSessionFileChange2 } from '../../../../workbench/contrib/chat/common/chatSessionsService.js';
-import { chatEditingWidgetFileStateContextKey, ModifiedFileEntryState } from '../../../../workbench/contrib/chat/common/editing/chatEditingService.js';
 import { createFileIconThemableTreeContainerScope } from '../../../../workbench/contrib/files/browser/views/explorerView.js';
 import { ACTIVE_GROUP, IEditorService, SIDE_GROUP } from '../../../../workbench/services/editor/common/editorService.js';
 import { IExtensionService } from '../../../../workbench/services/extensions/common/extensions.js';
 import { IWorkbenchLayoutService } from '../../../../workbench/services/layout/browser/layoutService.js';
 import { ISessionsManagementService } from '../../sessions/browser/sessionsManagementService.js';
 import { CodeReviewStateKind, getCodeReviewFilesFromSessionChanges, getCodeReviewVersion, ICodeReviewService, PRReviewStateKind } from '../../codeReview/browser/codeReviewService.js';
-import { IAgentFeedbackService } from '../../agentFeedback/browser/agentFeedbackService.js';
-import { GitDiffChange, GitRepositoryState, IGitRepository, IGitService } from '../../../../workbench/contrib/git/common/gitService.js';
 import { CIStatusWidget } from './checksWidget.js';
-import { arrayEqualsC, structuralEquals } from '../../../../base/common/equals.js';
 import { GITHUB_REMOTE_FILE_SCHEME, SessionStatus } from '../../sessions/common/sessionData.js';
 import { Orientation } from '../../../../base/browser/ui/sash/sash.js';
 import { IView, Sizing, SplitView } from '../../../../base/browser/ui/splitview/splitview.js';
@@ -76,569 +66,15 @@ import { EditorResourceAccessor, SideBySideEditor } from '../../../../workbench/
 import { logChangesViewFileSelect, logChangesViewVersionModeChange, logChangesViewViewModeChange } from '../../../common/sessionsTelemetry.js';
 import { ChecksViewModel } from './checksViewModel.js';
 import { ActiveSessionContextKeys, CHANGES_VIEW_CONTAINER_ID, CHANGES_VIEW_ID, ChangesContextKeys, ChangesVersionMode, ChangesViewMode, IsolationMode } from '../common/changes.js';
+import { buildTreeChildren, ChangesTreeElement, ChangesTreeRenderer, IChangesFileItem, IChangesTreeRootInfo, isChangesFileItem, toIChangesFileItem } from './changesViewRenderer.js';
+import { ChangesViewModel } from './changesViewModel.js';
+import { ResourceTree } from '../../../../base/common/resourceTree.js';
 
 const $ = dom.$;
 
 // --- Constants
 
 const RUN_SESSION_CODE_REVIEW_ACTION_ID = 'sessions.codeReview.run';
-
-// --- List Item
-
-type ChangeType = 'added' | 'modified' | 'deleted' | 'none';
-
-interface IChangesFileItem {
-	readonly type: 'file';
-	readonly uri: URI;
-	readonly originalUri?: URI;
-	readonly state: ModifiedFileEntryState;
-	readonly isDeletion: boolean;
-	readonly changeType: ChangeType;
-	readonly linesAdded: number;
-	readonly linesRemoved: number;
-}
-
-interface IChangesRootItem {
-	readonly type: 'root';
-	readonly uri: URI;
-	readonly name: string;
-}
-
-interface IChangesTreeRootInfo {
-	readonly root: IChangesRootItem;
-	readonly resourceTreeRootUri: URI;
-}
-
-type ChangesTreeElement = IChangesRootItem | IChangesFileItem | IResourceNode<IChangesFileItem, undefined>;
-
-function isChangesFileItem(element: ChangesTreeElement): element is IChangesFileItem {
-	return !ResourceTree.isResourceNode(element) && element.type === 'file';
-}
-
-function isChangesRootItem(element: ChangesTreeElement): element is IChangesRootItem {
-	return !ResourceTree.isResourceNode(element) && element.type === 'root';
-}
-
-/**
- * Builds a tree of `ICompressedTreeElement<ChangesTreeElement>` from a flat list of file items
- * using a `ResourceTree` to group files by their directory path segments.
- */
-function buildTreeChildren(items: IChangesFileItem[], treeRootInfo?: IChangesTreeRootInfo): ICompressedTreeElement<ChangesTreeElement>[] {
-	if (items.length === 0) {
-		return [];
-	}
-
-	let rootUri = treeRootInfo?.resourceTreeRootUri ?? URI.file('/');
-
-	// For github-remote-file URIs, set the root to /{owner}/{repo}/{ref}
-	// so the tree shows repo-relative paths instead of internal URI segments.
-	if (!treeRootInfo && items[0].uri.scheme === GITHUB_REMOTE_FILE_SCHEME) {
-		const parts = items[0].uri.path.split('/').filter(Boolean);
-		if (parts.length >= 3) {
-			rootUri = items[0].uri.with({ path: '/' + parts.slice(0, 3).join('/') });
-		}
-	}
-
-	const resourceTree = new ResourceTree<IChangesFileItem, undefined>(undefined, rootUri, extUriBiasedIgnorePathCase);
-	for (const item of items) {
-		resourceTree.add(item.uri, item);
-	}
-
-	function convertChildren(parent: IResourceNode<IChangesFileItem, undefined>): ICompressedTreeElement<ChangesTreeElement>[] {
-		const result: ICompressedTreeElement<ChangesTreeElement>[] = [];
-		for (const child of parent.children) {
-			if (child.element && child.childrenCount === 0) {
-				// Leaf node — just the file item
-				result.push({
-					element: child.element,
-					collapsible: false,
-					incompressible: true,
-				});
-			} else {
-				// Folder node. Ensure that the first level of folders under
-				// the root folder are not being collapsed with the root folder
-				// as that is a special node showing the workspace folder and
-				// branch information.
-				result.push({
-					element: child,
-					children: convertChildren(child),
-					incompressible: parent === resourceTree.root,
-					collapsible: true,
-					collapsed: false,
-				});
-			}
-		}
-		return result;
-	}
-
-	const children = convertChildren(resourceTree.root);
-	if (!treeRootInfo) {
-		return children;
-	}
-
-	return [{
-		element: treeRootInfo.root,
-		children,
-		collapsible: true,
-		collapsed: false,
-		incompressible: true,
-	}];
-}
-
-function toIChatSessionFileChange2(changes: GitDiffChange[], modifiedRef: string | undefined, originalRef: string | undefined): IChatSessionFileChange2[] {
-	return changes.map(change => ({
-		uri: change.uri,
-		originalUri: change.originalUri
-			? originalRef
-				? change.originalUri.with({ scheme: 'git', query: JSON.stringify({ path: change.originalUri.fsPath, ref: originalRef }) })
-				: change.originalUri
-			: undefined,
-		modifiedUri: change.modifiedUri
-			? modifiedRef
-				? change.modifiedUri.with({ scheme: 'git', query: JSON.stringify({ path: change.modifiedUri.fsPath, ref: modifiedRef }) })
-				: change.modifiedUri
-			: undefined,
-		insertions: change.insertions,
-		deletions: change.deletions,
-	} satisfies IChatSessionFileChange2));
-}
-
-function toIChangesFileItem(changes: readonly (IChatSessionFileChange | IChatSessionFileChange2)[]): IChangesFileItem[] {
-	return changes.map(change => {
-		const isAddition = change.originalUri === undefined;
-		const isDeletion = change.modifiedUri === undefined;
-		const uri = isIChatSessionFileChange2(change)
-			? change.uri
-			: change.modifiedUri;
-
-		return {
-			type: 'file',
-			uri,
-			originalUri: change.originalUri,
-			isDeletion,
-			state: ModifiedFileEntryState.Accepted,
-			changeType: isAddition
-				? 'added'
-				: isDeletion
-					? 'deleted'
-					: 'modified',
-			linesAdded: change.insertions,
-			linesRemoved: change.deletions
-		} satisfies IChangesFileItem;
-	});
-}
-
-// --- View Model
-
-interface ActiveSessionState {
-	readonly isolationMode: IsolationMode;
-	readonly hasGitRepository: boolean;
-	readonly branchName: string | undefined;
-	readonly baseBranchName: string | undefined;
-	readonly isMergeBaseBranchProtected: boolean | undefined;
-	readonly incomingChanges: number | undefined;
-	readonly outgoingChanges: number | undefined;
-	readonly uncommittedChanges: number | undefined;
-	readonly hasPullRequest: boolean | undefined;
-	readonly hasOpenPullRequest: boolean | undefined;
-}
-
-class ChangesViewModel extends Disposable {
-	readonly activeSessionResourceObs: IObservable<URI | undefined>;
-	readonly activeSessionRepositoryObs: IObservableWithChange<IGitRepository | undefined>;
-	readonly activeSessionRepositoryStateObs: IObservableWithChange<GitRepositoryState | undefined>;
-	readonly activeSessionChangesObs: IObservable<readonly (IChatSessionFileChange | IChatSessionFileChange2)[]>;
-	readonly activeSessionHasGitRepositoryObs: IObservable<boolean>;
-	readonly activeSessionFirstCheckpointRefObs: IObservable<string | undefined>;
-	readonly activeSessionLastCheckpointRefObs: IObservable<string | undefined>;
-	readonly activeSessionReviewCommentCountByFileObs: IObservable<Map<string, number>>;
-	readonly activeSessionAgentFeedbackCountByFileObs: IObservable<Map<string, number>>;
-	readonly activeSessionStateObs: IObservable<ActiveSessionState | undefined>;
-	readonly activeSessionIsLoadingObs: IObservable<boolean>;
-
-	private _activeSessionMetadataObs!: IObservable<{ readonly [key: string]: unknown } | undefined>;
-	private _activeSessionAllChangesPromiseObs!: IObservableWithChange<IObservable<GitDiffChange[] | undefined>>;
-	private _activeSessionLastTurnChangesPromiseObs!: IObservableWithChange<IObservable<GitDiffChange[] | undefined>>;
-
-	readonly versionModeObs: ISettableObservable<ChangesVersionMode>;
-	setVersionMode(mode: ChangesVersionMode): void {
-		if (this.versionModeObs.get() === mode) {
-			return;
-		}
-		this.versionModeObs.set(mode, undefined);
-	}
-
-	readonly viewModeObs: ISettableObservable<ChangesViewMode>;
-	setViewMode(mode: ChangesViewMode): void {
-		if (this.viewModeObs.get() === mode) {
-			return;
-		}
-		this.viewModeObs.set(mode, undefined);
-		this.storageService.store('changesView.viewMode', mode, StorageScope.WORKSPACE, StorageTarget.USER);
-	}
-
-	constructor(
-		@IAgentFeedbackService private readonly agentFeedbackService: IAgentFeedbackService,
-		@IAgentSessionsService private readonly agentSessionsService: IAgentSessionsService,
-		@ICodeReviewService private readonly codeReviewService: ICodeReviewService,
-		@IGitService private readonly gitService: IGitService,
-		@ISessionsManagementService private readonly sessionManagementService: ISessionsManagementService,
-		@IStorageService private readonly storageService: IStorageService,
-	) {
-		super();
-
-		// Active session resource
-		this.activeSessionResourceObs = derivedOpts({ equalsFn: isEqual }, reader => {
-			const activeSession = this.sessionManagementService.activeSession.read(reader);
-			return activeSession?.resource;
-		});
-
-		// Active session metadata
-		this._activeSessionMetadataObs = this._getActiveSessionMetadata();
-
-		// Active session has git repository
-		this.activeSessionHasGitRepositoryObs = derived(reader => {
-			const metadata = this._activeSessionMetadataObs.read(reader);
-			return metadata?.repositoryPath !== undefined;
-		});
-
-		// Active session first checkpoint ref
-		this.activeSessionFirstCheckpointRefObs = derived(reader => {
-			const metadata = this._activeSessionMetadataObs.read(reader);
-			return metadata?.firstCheckpointRef as string | undefined;
-		});
-
-		// Active session last checkpoint ref
-		this.activeSessionLastCheckpointRefObs = derived(reader => {
-			const metadata = this._activeSessionMetadataObs.read(reader);
-			return metadata?.lastCheckpointRef as string | undefined;
-		});
-
-		// Active session repository
-		const { repository, repositoryState } = this._getActiveSessionGitRepository();
-		this.activeSessionRepositoryObs = repository;
-		this.activeSessionRepositoryStateObs = repositoryState;
-
-		// Active session state
-		const { isLoading, state } = this._getActiveSessionState();
-		this.activeSessionIsLoadingObs = isLoading;
-		this.activeSessionStateObs = state;
-
-		// Active session changes
-		this.activeSessionChangesObs = this._getActiveSessionChanges();
-
-		// Active session review comment count by file
-		this.activeSessionReviewCommentCountByFileObs = this._getActiveSessionReviewComments();
-
-		// Active session agent feedback count by file
-		this.activeSessionAgentFeedbackCountByFileObs = this._getActiveSessionAgentFeedback();
-
-		// Version mode
-		this.versionModeObs = observableValue<ChangesVersionMode>(this, ChangesVersionMode.BranchChanges);
-
-		this._register(runOnChange(this.activeSessionResourceObs, () => {
-			this.setVersionMode(ChangesVersionMode.BranchChanges);
-		}));
-
-		// View mode
-		const storedMode = this.storageService.get('changesView.viewMode', StorageScope.WORKSPACE);
-		const initialMode = storedMode === ChangesViewMode.Tree ? ChangesViewMode.Tree : ChangesViewMode.List;
-		this.viewModeObs = observableValue<ChangesViewMode>(this, initialMode);
-	}
-
-	private _getActiveSessionMetadata(): IObservable<{ readonly [key: string]: unknown } | undefined> {
-		const sessionsChangedSignal = observableSignalFromEvent(this,
-			this.sessionManagementService.onDidChangeSessions);
-
-		return derivedOpts<{ readonly [key: string]: unknown } | undefined>({
-			equalsFn: structuralEquals
-		}, reader => {
-			const sessionResource = this.activeSessionResourceObs.read(reader);
-			if (!sessionResource) {
-				return undefined;
-			}
-
-			sessionsChangedSignal.read(reader);
-			const model = this.agentSessionsService.getSession(sessionResource);
-			return model?.metadata;
-		});
-	}
-
-	private _getActiveSessionGitRepository(): { repository: IObservable<IGitRepository | undefined>; repositoryState: IObservable<GitRepositoryState | undefined> } {
-		const activeSessionRepositoryObs = derived(reader => {
-			const activeSession = this.sessionManagementService.activeSession.read(reader);
-			return activeSession?.workspace.read(reader)?.repositories[0];
-		});
-
-		const activeSessionRepositoryPromiseObs = derived(reader => {
-			const activeSessionResource = this.activeSessionResourceObs.read(reader);
-			if (!activeSessionResource) {
-				return constObservable(undefined);
-			}
-
-			const activeSessionRepository = activeSessionRepositoryObs.read(reader);
-			const workingDirectory = activeSessionRepository?.workingDirectory ?? activeSessionRepository?.uri;
-			if (!workingDirectory) {
-				return constObservable(undefined);
-			}
-
-			return new ObservablePromise(this.gitService.openRepository(workingDirectory)).resolvedValue;
-		});
-
-		const activeSessionGitRepositoryObs = derived<IGitRepository | undefined>(reader => {
-			const activeSessionRepositoryPromise = activeSessionRepositoryPromiseObs.read(reader);
-			if (activeSessionRepositoryPromise === undefined) {
-				return undefined;
-			}
-
-			return activeSessionRepositoryPromise.read(reader);
-		});
-
-		const activeSessionGitRepositoryStateObs = derived(reader => {
-			const repository = activeSessionGitRepositoryObs.read(reader);
-			const repositoryState = repository?.state.read(reader);
-
-			// If the repository has no HEAD, it is likely not fully loaded yet.
-			// Treat it as undefined to avoid showing incorrect information to
-			// the user.
-			if (!repositoryState?.HEAD) {
-				return undefined;
-			}
-
-			return repositoryState;
-		});
-
-		return {
-			repository: activeSessionGitRepositoryObs,
-			repositoryState: activeSessionGitRepositoryStateObs
-		};
-	}
-
-	private _getActiveSessionChanges(): IObservable<readonly (IChatSessionFileChange | IChatSessionFileChange2)[]> {
-		// Changes
-		const activeSessionChangesObs = derived(reader => {
-			const activeSession = this.sessionManagementService.activeSession.read(reader);
-			if (!activeSession) {
-				return Iterable.empty();
-			}
-			return activeSession.changes.read(reader);
-		});
-
-		// All changes
-		this._activeSessionAllChangesPromiseObs = derived(reader => {
-			const repository = this.activeSessionRepositoryObs.read(reader);
-			const firstCheckpointRef = this.activeSessionFirstCheckpointRefObs.read(reader);
-			const lastCheckpointRef = this.activeSessionLastCheckpointRefObs.read(reader);
-
-			if (!repository || !firstCheckpointRef || !lastCheckpointRef) {
-				return constObservable([]);
-			}
-
-			const diffPromise = repository.diffBetweenWithStats(firstCheckpointRef, lastCheckpointRef);
-			return new ObservablePromise(diffPromise).resolvedValue;
-		});
-
-		const activeSessionAllChangesObs = derived(reader => {
-			const diffChanges = this._activeSessionAllChangesPromiseObs.read(reader).read(reader) ?? [];
-			const firstCheckpointRef = this.activeSessionFirstCheckpointRefObs.read(undefined);
-			const lastCheckpointRef = this.activeSessionLastCheckpointRefObs.read(undefined);
-
-			return toIChatSessionFileChange2(diffChanges, lastCheckpointRef, firstCheckpointRef);
-		});
-
-		// Last turn changes
-		this._activeSessionLastTurnChangesPromiseObs = derived(reader => {
-			const repository = this.activeSessionRepositoryObs.read(reader);
-			const lastCheckpointRef = this.activeSessionLastCheckpointRefObs.read(reader);
-
-			if (!repository || !lastCheckpointRef) {
-				return constObservable([]);
-			}
-
-			const diffPromise = repository.diffBetweenWithStats(`${lastCheckpointRef}^`, lastCheckpointRef);
-			return new ObservablePromise(diffPromise).resolvedValue;
-		});
-
-		const activeSessionLastTurnChangesObs = derived(reader => {
-			const diffChanges = this._activeSessionLastTurnChangesPromiseObs.read(reader).read(reader) ?? [];
-			const lastCheckpointRef = this.activeSessionLastCheckpointRefObs.read(undefined);
-
-			return toIChatSessionFileChange2(diffChanges, lastCheckpointRef, lastCheckpointRef ? `${lastCheckpointRef}^` : undefined);
-		});
-
-		return derivedOpts({
-			equalsFn: arrayEqualsC<IChatSessionFileChange | IChatSessionFileChange2>()
-		}, reader => {
-			const hasGitRepository = this.activeSessionHasGitRepositoryObs.read(reader);
-			if (!hasGitRepository) {
-				return [];
-			}
-
-			const versionMode = this.versionModeObs.read(reader);
-			if (versionMode === ChangesVersionMode.BranchChanges) {
-				return activeSessionChangesObs.read(reader);
-			} else if (versionMode === ChangesVersionMode.AllChanges) {
-				return activeSessionAllChangesObs.read(reader);
-			} else if (versionMode === ChangesVersionMode.LastTurn) {
-				return activeSessionLastTurnChangesObs.read(reader);
-			}
-
-			return [];
-		});
-	}
-
-	private _getActiveSessionState(): { isLoading: IObservable<boolean>; state: IObservable<ActiveSessionState | undefined> } {
-		const isLoadingObs = derived(reader => {
-			// If there is a git repository, wait for the repository to be opened first,
-			// as there are many context keys that depend on the repository information.
-			const hasGitRepository = this.activeSessionHasGitRepositoryObs.read(reader);
-			if (hasGitRepository && this.activeSessionRepositoryStateObs.read(reader) === undefined) {
-				return true;
-			}
-
-			// Branch changes
-			const versionMode = this.versionModeObs.read(reader);
-			if (versionMode === ChangesVersionMode.BranchChanges) {
-				return false;
-			}
-
-			// All changes
-			if (versionMode === ChangesVersionMode.AllChanges) {
-				const allChangesResult = this._activeSessionAllChangesPromiseObs.read(reader).read(reader);
-				return allChangesResult === undefined;
-			}
-
-			// Last turn changes
-			if (versionMode === ChangesVersionMode.LastTurn) {
-				const lastTurnChangesResult = this._activeSessionLastTurnChangesPromiseObs.read(reader).read(reader);
-				return lastTurnChangesResult === undefined;
-			}
-
-			return false;
-		});
-
-		const activeSessionStateObs = derivedObservableWithCache<ActiveSessionState | undefined>(this, (reader, lastValue) => {
-			const isLoading = isLoadingObs.read(reader);
-			const activeSession = this.sessionManagementService.activeSession.read(reader);
-			const repositoryState = this.activeSessionRepositoryStateObs.read(reader);
-			if (isLoading && repositoryState === undefined) {
-				return lastValue;
-			}
-
-			// Session state
-			const sessionMetadata = this._activeSessionMetadataObs.read(reader);
-			const workspace = activeSession?.workspace.read(reader);
-			const workspaceRepository = workspace?.repositories[0];
-			const hasGitRepository = this.activeSessionHasGitRepositoryObs.read(reader);
-			const branchName = sessionMetadata?.branchName as string | undefined;
-			const baseBranchName = sessionMetadata?.baseBranchName as string | undefined;
-			const isMergeBaseBranchProtected = workspaceRepository?.baseBranchProtected;
-			const isolationMode = workspaceRepository?.workingDirectory === undefined
-				? IsolationMode.Workspace
-				: IsolationMode.Worktree;
-
-			// Pull request state
-			const gitHubInfo = activeSession?.gitHubInfo.read(reader);
-			const hasPullRequest = gitHubInfo?.pullRequest?.uri !== undefined;
-			const hasOpenPullRequest = hasPullRequest &&
-				(gitHubInfo.pullRequest.icon?.id === Codicon.gitPullRequestDraft.id ||
-					gitHubInfo.pullRequest.icon?.id === Codicon.gitPullRequest.id);
-
-			// Repository state
-			const incomingChanges = hasGitRepository
-				? repositoryState?.HEAD?.behind ?? 0
-				: undefined;
-			const outgoingChanges = hasGitRepository
-				? repositoryState?.HEAD?.ahead ?? 0
-				: undefined;
-			const uncommittedChanges = hasGitRepository
-				? (repositoryState?.mergeChanges.length ?? 0) +
-				(repositoryState?.indexChanges.length ?? 0) +
-				(repositoryState?.workingTreeChanges.length ?? 0) +
-				(repositoryState?.untrackedChanges.length ?? 0)
-				: undefined;
-
-			return {
-				isolationMode,
-				hasGitRepository,
-				branchName,
-				baseBranchName,
-				isMergeBaseBranchProtected,
-				incomingChanges,
-				outgoingChanges,
-				uncommittedChanges,
-				hasPullRequest,
-				hasOpenPullRequest
-			} satisfies ActiveSessionState;
-		});
-
-		return {
-			isLoading: isLoadingObs,
-			state: derivedOpts({ equalsFn: structuralEquals },
-				reader => activeSessionStateObs.read(reader))
-		};
-	}
-
-	private _getActiveSessionReviewComments(): IObservable<Map<string, number>> {
-		return derived(reader => {
-			const sessionResource = this.activeSessionResourceObs.read(reader);
-			const changes = [...this.activeSessionChangesObs.read(reader)];
-
-			if (!sessionResource) {
-				return new Map<string, number>();
-			}
-
-			const result = new Map<string, number>();
-			const prReviewState = this.codeReviewService.getPRReviewState(sessionResource).read(reader);
-			if (prReviewState.kind === PRReviewStateKind.Loaded) {
-				for (const comment of prReviewState.comments) {
-					const uriKey = comment.uri.fsPath;
-					result.set(uriKey, (result.get(uriKey) ?? 0) + 1);
-				}
-			}
-
-			if (changes.length === 0) {
-				return result;
-			}
-
-			const reviewFiles = getCodeReviewFilesFromSessionChanges(changes);
-			const reviewVersion = getCodeReviewVersion(reviewFiles);
-			const reviewState = this.codeReviewService.getReviewState(sessionResource).read(reader);
-
-			if (reviewState.kind !== CodeReviewStateKind.Result || reviewState.version !== reviewVersion) {
-				return result;
-			}
-
-			for (const comment of reviewState.comments) {
-				const uriKey = comment.uri.fsPath;
-				result.set(uriKey, (result.get(uriKey) ?? 0) + 1);
-			}
-
-			return result;
-		});
-	}
-
-	private _getActiveSessionAgentFeedback(): IObservable<Map<string, number>> {
-		return derived(reader => {
-			const sessionResource = this.activeSessionResourceObs.read(reader);
-			if (!sessionResource) {
-				return new Map<string, number>();
-			}
-
-			observableSignalFromEvent(this, this.agentFeedbackService.onDidChangeFeedback).read(reader);
-
-			const feedbackItems = this.agentFeedbackService.getFeedback(sessionResource);
-			const result = new Map<string, number>();
-			for (const item of feedbackItems) {
-				if (!item.sourcePRReviewCommentId) {
-					const uriKey = item.resourceUri.fsPath;
-					result.set(uriKey, (result.get(uriKey) ?? 0) + 1);
-				}
-			}
-			return result;
-		});
-	}
-}
 
 // --- View Pane
 
@@ -1520,7 +956,7 @@ class ChangesViewActionRunner extends ActionRunner {
 	}
 }
 
-// --- Tree Delegate & Renderer
+// --- Tree Delegate
 
 class ChangesTreeDelegate implements IListVirtualDelegate<ChangesTreeElement> {
 	static readonly ROW_HEIGHT = 22;
@@ -1531,274 +967,6 @@ class ChangesTreeDelegate implements IListVirtualDelegate<ChangesTreeElement> {
 
 	getTemplateId(_element: ChangesTreeElement): string {
 		return ChangesTreeRenderer.TEMPLATE_ID;
-	}
-}
-
-interface IChangesTreeTemplate {
-	readonly label: IResourceLabel;
-	readonly toolbar: MenuWorkbenchToolBar | undefined;
-	readonly contextKeyService: IContextKeyService | undefined;
-	readonly reviewCommentsBadge: HTMLElement;
-	readonly agentFeedbackBadge: HTMLElement;
-	readonly decorationBadge: HTMLElement;
-	readonly addedSpan: HTMLElement;
-	readonly removedSpan: HTMLElement;
-	readonly lineCountsContainer: HTMLElement;
-	readonly elementDisposables: DisposableStore;
-	readonly templateDisposables: DisposableStore;
-}
-
-class ChangesTreeRenderer implements ICompressibleTreeRenderer<ChangesTreeElement, void, IChangesTreeTemplate> {
-	static TEMPLATE_ID = 'changesTreeRenderer';
-	readonly templateId: string = ChangesTreeRenderer.TEMPLATE_ID;
-
-	constructor(
-		private viewModel: ChangesViewModel,
-		private labels: ResourceLabels,
-		private actionRunner: ActionRunner | undefined,
-		private getRootUri: () => URI | undefined,
-		@IInstantiationService private readonly instantiationService: IInstantiationService,
-		@IContextKeyService private readonly contextKeyService: IContextKeyService,
-		@ILabelService private readonly labelService: ILabelService,
-		@ISessionsManagementService private readonly sessionManagementService: ISessionsManagementService,
-	) { }
-
-	renderTemplate(container: HTMLElement): IChangesTreeTemplate {
-		const templateDisposables = new DisposableStore();
-		const label = templateDisposables.add(this.labels.create(container, { supportHighlights: true, supportIcons: true }));
-
-		const reviewCommentsBadge = dom.$('.changes-review-comments-badge');
-		label.element.appendChild(reviewCommentsBadge);
-
-		const agentFeedbackBadge = dom.$('.changes-agent-feedback-badge');
-		label.element.appendChild(agentFeedbackBadge);
-
-		const lineCountsContainer = $('.working-set-line-counts');
-		const addedSpan = dom.$('.working-set-lines-added');
-		const removedSpan = dom.$('.working-set-lines-removed');
-		lineCountsContainer.appendChild(addedSpan);
-		lineCountsContainer.appendChild(removedSpan);
-		label.element.appendChild(lineCountsContainer);
-
-		const actionBarContainer = $('.chat-collapsible-list-action-bar');
-		const contextKeyService = templateDisposables.add(this.contextKeyService.createScoped(actionBarContainer));
-		const scopedInstantiationService = templateDisposables.add(this.instantiationService.createChild(new ServiceCollection([IContextKeyService, contextKeyService])));
-		const toolbar = templateDisposables.add(scopedInstantiationService.createInstance(MenuWorkbenchToolBar, actionBarContainer, MenuId.ChatEditingSessionChangeToolbar, { menuOptions: { shouldForwardArgs: true, arg: undefined }, actionRunner: this.actionRunner }));
-		label.element.appendChild(actionBarContainer);
-
-		templateDisposables.add(bindContextKey(ChatContextKeys.agentSessionType, contextKeyService, reader => {
-			const activeSession = this.sessionManagementService.activeSession.read(reader);
-			return activeSession?.sessionType ?? '';
-		}));
-
-		templateDisposables.add(bindContextKey(ActiveSessionContextKeys.HasGitRepository, contextKeyService, reader => {
-			return this.viewModel.activeSessionHasGitRepositoryObs.read(reader);
-		}));
-
-		templateDisposables.add(bindContextKey(ChangesContextKeys.VersionMode, contextKeyService, reader => {
-			return this.viewModel.versionModeObs.read(reader);
-		}));
-
-		const decorationBadge = dom.$('.changes-decoration-badge');
-		label.element.appendChild(decorationBadge);
-
-		return { label, toolbar, contextKeyService, reviewCommentsBadge, agentFeedbackBadge, decorationBadge, addedSpan, removedSpan, lineCountsContainer, elementDisposables: new DisposableStore(), templateDisposables };
-	}
-
-	renderElement(node: ITreeNode<ChangesTreeElement, void>, _index: number, templateData: IChangesTreeTemplate): void {
-		const element = node.element;
-		templateData.label.element.style.display = 'flex';
-
-		if (isChangesRootItem(element)) {
-			// Root element
-			this.renderRootElement(element, templateData);
-		} else if (ResourceTree.isResourceNode(element)) {
-			// Folder element
-			this.renderFolderElement(element, templateData);
-		} else {
-			// File element
-			this.renderFileElement(element, templateData);
-		}
-	}
-
-	renderCompressedElements(node: ITreeNode<ICompressedTreeNode<ChangesTreeElement>, void>, _index: number, templateData: IChangesTreeTemplate): void {
-		const compressed = node.element as ICompressedTreeNode<IResourceNode<IChangesFileItem, undefined>>;
-		const folder = compressed.elements[compressed.elements.length - 1];
-
-		templateData.label.element.style.display = 'flex';
-
-		const label = compressed.elements.map(e => e.name);
-		templateData.label.setResource({ resource: folder.uri, name: label }, {
-			fileKind: FileKind.FOLDER,
-			separator: this.labelService.getSeparator(folder.uri.scheme),
-		});
-
-		// Hide file-specific decorations for folders
-		templateData.reviewCommentsBadge.style.display = 'none';
-		templateData.agentFeedbackBadge.style.display = 'none';
-		templateData.decorationBadge.style.display = 'none';
-		templateData.lineCountsContainer.style.display = 'none';
-
-		if (templateData.toolbar) {
-			templateData.toolbar.context = folder;
-		}
-		if (templateData.contextKeyService) {
-			chatEditingWidgetFileStateContextKey.bindTo(templateData.contextKeyService).set(undefined!);
-		}
-	}
-
-	private renderFileElement(data: IChangesFileItem, templateData: IChangesTreeTemplate): void {
-		const root = this.getRootUri();
-		const viewMode = this.viewModel.viewModeObs.get();
-
-		templateData.label.setResource({
-			resource: data.uri,
-			name: basename(data.uri.path),
-			description: viewMode === ChangesViewMode.List
-				? root
-					? relativePath(root, dirname(data.uri))
-					: undefined
-				: undefined,
-		}, {
-			fileKind: FileKind.FILE,
-			fileDecorations: undefined,
-			strikethrough: data.changeType === 'deleted'
-		});
-
-		const showChangeDecorations = data.changeType !== 'none';
-
-		// Show file-specific decorations for changed files only
-		templateData.lineCountsContainer.style.display = showChangeDecorations ? '' : 'none';
-		templateData.decorationBadge.style.display = showChangeDecorations ? '' : 'none';
-
-		// Review comments
-		templateData.elementDisposables.add(autorun(reader => {
-			const reviewCommentByFile = this.viewModel.activeSessionReviewCommentCountByFileObs.read(reader);
-			const reviewCommentCount = reviewCommentByFile?.get(data.uri.fsPath) ?? 0;
-
-			if (reviewCommentCount > 0) {
-				templateData.reviewCommentsBadge.style.display = '';
-				templateData.reviewCommentsBadge.className = 'changes-review-comments-badge';
-				templateData.reviewCommentsBadge.replaceChildren(
-					dom.$('.codicon.codicon-comment-unresolved'),
-					dom.$('span', undefined, `${reviewCommentCount}`)
-				);
-			} else {
-				templateData.reviewCommentsBadge.style.display = 'none';
-				templateData.reviewCommentsBadge.replaceChildren();
-			}
-		}));
-
-		// Agent feedback
-		templateData.elementDisposables.add(autorun(reader => {
-			const agentFeedbackByFile = this.viewModel.activeSessionAgentFeedbackCountByFileObs.read(reader);
-			const agentFeedbackCount = agentFeedbackByFile?.get(data.uri.fsPath) ?? 0;
-
-			if (agentFeedbackCount > 0) {
-				templateData.agentFeedbackBadge.style.display = '';
-				templateData.agentFeedbackBadge.className = 'changes-agent-feedback-badge';
-				templateData.agentFeedbackBadge.replaceChildren(
-					dom.$('.codicon.codicon-comment'),
-					dom.$('span', undefined, `${agentFeedbackCount}`)
-				);
-			} else {
-				templateData.agentFeedbackBadge.style.display = 'none';
-				templateData.agentFeedbackBadge.replaceChildren();
-			}
-		}));
-
-		const badge = templateData.decorationBadge;
-		badge.className = 'changes-decoration-badge';
-		if (showChangeDecorations) {
-			// Update decoration badge (A/M/D)
-			switch (data.changeType) {
-				case 'added':
-					badge.textContent = 'A';
-					badge.classList.add('added');
-					break;
-				case 'deleted':
-					badge.textContent = 'D';
-					badge.classList.add('deleted');
-					break;
-				case 'modified':
-				default:
-					badge.textContent = 'M';
-					badge.classList.add('modified');
-					break;
-			}
-
-			templateData.addedSpan.textContent = `+${data.linesAdded}`;
-			templateData.removedSpan.textContent = `-${data.linesRemoved}`;
-
-			// eslint-disable-next-line no-restricted-syntax
-			templateData.label.element.querySelector('.monaco-icon-name-container')?.classList.add('modified');
-		} else {
-			badge.textContent = '';
-			// eslint-disable-next-line no-restricted-syntax
-			templateData.label.element.querySelector('.monaco-icon-name-container')?.classList.remove('modified');
-		}
-
-		if (templateData.toolbar) {
-			templateData.toolbar.context = data;
-		}
-		if (templateData.contextKeyService) {
-			chatEditingWidgetFileStateContextKey.bindTo(templateData.contextKeyService).set(data.state);
-		}
-	}
-
-	private renderRootElement(data: IChangesRootItem, templateData: IChangesTreeTemplate): void {
-		templateData.label.setResource({
-			resource: data.uri,
-			name: data.name,
-		}, {
-			fileKind: FileKind.ROOT_FOLDER,
-			separator: this.labelService.getSeparator(data.uri.scheme, data.uri.authority),
-		});
-
-		templateData.reviewCommentsBadge.style.display = 'none';
-		templateData.agentFeedbackBadge.style.display = 'none';
-		templateData.decorationBadge.style.display = 'none';
-		templateData.lineCountsContainer.style.display = 'none';
-
-		if (templateData.toolbar) {
-			templateData.toolbar.context = data.uri;
-		}
-		if (templateData.contextKeyService) {
-			chatEditingWidgetFileStateContextKey.bindTo(templateData.contextKeyService).set(undefined!);
-		}
-	}
-
-	private renderFolderElement(node: IResourceNode<IChangesFileItem, undefined>, templateData: IChangesTreeTemplate): void {
-		templateData.label.setFile(node.uri, {
-			fileKind: FileKind.FOLDER,
-			hidePath: true,
-		});
-
-		// Hide file-specific decorations for folders
-		templateData.reviewCommentsBadge.style.display = 'none';
-		templateData.agentFeedbackBadge.style.display = 'none';
-		templateData.decorationBadge.style.display = 'none';
-		templateData.lineCountsContainer.style.display = 'none';
-
-		if (templateData.toolbar) {
-			templateData.toolbar.context = node;
-		}
-		if (templateData.contextKeyService) {
-			chatEditingWidgetFileStateContextKey.bindTo(templateData.contextKeyService).set(undefined!);
-		}
-	}
-
-	disposeElement(_element: ITreeNode<ChangesTreeElement, void>, _index: number, templateData: IChangesTreeTemplate): void {
-		templateData.elementDisposables.clear();
-	}
-
-	disposeCompressedElements(_element: ITreeNode<ICompressedTreeNode<ChangesTreeElement>, void>, _index: number, templateData: IChangesTreeTemplate): void {
-		templateData.elementDisposables.clear();
-	}
-
-	disposeTemplate(templateData: IChangesTreeTemplate): void {
-		templateData.elementDisposables.dispose();
-		templateData.templateDisposables.dispose();
 	}
 }
